+++ app/main.py
@@ -7,9 +7,8 @@ from app.core.auth_util import User_ID_Dep
 from app.core.auth_util import init_firebase
 from app.core.sql_util import SessionDep
 from app.models.data_model import UserDocData
-from app.utils.file_util import validate_pdf_file
 # from app.utils.text_util import extract_text_from_pdf
-from app.utils.text_util import pypdf2_extract
+from app.utils.pdf_util import pypdf2_extract, validate_pdf_file
 
 
 @asynccontextmanager
@@ -17,13 +16,14 @@ async def lifespan(app: FastAPI):
     init_firebase()
     yield
 
+
 app = FastAPI(lifespan=lifespan,
               # dependencies=[Depends(verify_and_get_user)]
               )
 
 
 @app.get("/")
-async def root(session : SessionDep):
+async def root(session: SessionDep):
     print(session)
     return {"message": "Hello World"}
 
@@ -35,5 +35,4 @@ async def upload_file(file: UploadFile, user_id: User_ID_Dep):
     # print(user_id)
     user_doc_data = UserDocData(user_id=user_id, extracted_text=extracted_text, upload_time=datetime.now())
 
-    return {"filename": file.filename, "text": extracted_text, "user":user_id}
-    
+    return {"filename": file.filename, "text": extracted_text, "user": user_id}
