+++ app/utils/pdf_util.py
@@ -1,8 +1,10 @@
 import tempfile
 
-import pymupdf4llm
-from fastapi import UploadFile
 from PyPDF2 import PdfReader
+from fastapi import UploadFile
+
+from app.core.config import settings
+
 
 # def extract_text_from_pdf(upload_file: UploadFile) -> str:
 #     with tempfile.NamedTemporaryFile(delete=False) as temp_file:
@@ -23,3 +25,16 @@ def pypdf2_extract(upload_file: UploadFile) -> str:
                 all_text += text
 
     return all_text
+
+
+class FileValidationError(Exception):
+    """Custom exception for file validation errors."""
+    pass
+
+
+def validate_pdf_file(upload_file: UploadFile) -> None:
+    if not upload_file.filename.lower().endswith('.pdf'):
+        raise FileValidationError("File is not a PDF.")
+
+    if upload_file.size > settings.max_file_size:
+        raise FileValidationError("File size exceeds 5 MB.")
